[BugFix] Fixed MenuCrewMember _unitIdx attribute type

+++ src/interface/menus/MenuCrewMember.cc
@@ -76,7 +76,7 @@ void MenuCrewMember::getOut()
   game::Status::setStateAttributes(
     std::make_shared<Coords> (vehicle_coords),
     std::make_shared<e_unit_role> (_role),
-    std::make_shared<int> (_unitIdx)
+    std::make_shared<size_t> (_unitIdx)
   );
 
   game::Status::resumeState();
